refactor: Split FileListHeader file
